I15-852 (9.30) Default DataCollectionGroup ExperimentType to OSC
Also, add a detector check, since the dataCollectionGroup will only be
used for single crystal experiments, and these experiments always
require cbf data, so only a pilatus detector with a cbf datawriter is
suitable.

Run the check both before and after the context yield so that the
warning is more likely to be seen after the context if it wasn't seen at
the start.
Change-Id: Ic136cec8d46a7e8819525fca537f64eb50320a41

Change-Id: I14097880ec7b0a1b672c0038998ee279482ecfc4