I15-718 Ensure that EpicsRockingScannable doesn't fail at end of scan
Moving back to start at the end of scan gets the position and then just
moves to that position. Really it should get the position and record
only the input names (stripping off the extra name fields). This works
around the problem by trying the move again with the 0th element if the
full position causes an exception.

Change-Id: I011156bc919f9245baac71a2d80eb3ef26092489