I18 : Add Medipix detector to server and client
An ADDetector object has been created for Medipix/Merlin detector for
use with the XES spectrometer; frames are collected during during step
scans using internal triggering. 'Collection strategy' decorators are
used to restore the trigger mode, acquire time and image mode are at the
end of a scan.
A 'live stream view' for the detector has also been added to allow the
live data to be viewed, start and stop the detector and set the ROI.

Change-Id: I572cd1d2d101907305378f15979039484cfe4b16