K11: Select correct XY shape when loading diffraction tomography scan
ShapeControls attempts to match the TrajectoryShape of the first traj-
ectory in the scan with the ones it has been initialised with. However
for diffraction tomography, the first trajectory is X-theta so the check
does not produce anything meaningful. This change lets a Function be
injected to ShapeControls to result in a sensible shape for the given
scan.

Change-Id: Ibb1b758cd11d4353d170e3ab06e3e40c521346bb