B18 : Avoid class cast exception when moving sample wheel
A check has been added to B18SampleEnvironmentIterator to ensure the
sample wheel scannable is the correct type before moving it using filter name.
This avoids a class cast exception when attempting to do this in dummy
mode (since the scannable is the wrong type) and improves the error message
reporting by throwing a meaningful exception.

Change-Id: I863ddfd3e10b57591edeb0f59bc79382387ffb4f