I16 (9.33) Make samples easier to find, tidy up & update baseline
Add implicit properties to the beam bean to make NXsample contributions
easier to find.

Convert comments of the form <-- ... --> that may need to be nested into
comments of the form <?ignore ... ?> to facility that.

Remove comment only lines from NXmx.yaml

Use the most recent I16_634 scan as it's baseline.

Change-Id: Ia62d9e96d31c3f330af802a501591385722a3e30