B18 : Updates for continuous scans
- A new BufferedScannablePositions class has been created. This is an
implementation of BufferedDetector that stores the position of a
scannable at regularly spaced time intervals. The list of positions
stored is returned by the readout() method.
- continuous_scans.py has been updated to the BufferedScannablePositions
to record the position of the moving scannable. This ensures that the
position of the scannable is measured at the same time as the detector
measurement, and improves the accuracy of the axis values.
Other improvemeents :
- Scans now work correctly when running in negative direction
- Scans can run bi-directionally by passing bidirectional=True to
createContinuousScan method.
- Detector list is created in safe way to avoid accidentally sharing
the default value between method invocations.

Change-Id: I06839e1c73308e661e9e1ecc497684865b5418fb