I20-629 : Rationalise XES scan setup
- All XES scans are now set up in the same way : by determining the
moving and stationary scannables and the list of energies/times to be
used and assembling a complete list of scan arguments.
This method has the benefit that all output and detector preparer
objects have access to the original scan settings bean, and so can
determine unambiguously what type of scan is being done.
- XES unit tests have been updated
- Nexus services are now set in the mono optimisation test so that
tests can be run successfully from Eclipse.

Change-Id: I22f1fbdb84c2c52c4ec9f03dd17a01bcf47ccaee