I18 : Add Andor detector to server configuration
Andor detector has been added to the server layout, using a fairly
standard area detector configuration, similar to the Medipix.
The Hdf file writer plugin is used to store image data (the images are
large and cause a 'max array bytes' channel access error if the NDArray
plugin is used to read and record the data directly in the Nexus file).

Change-Id: Ia1aeab7e759f1b6affc3d2d3eff163647c8ab97d