I18 : Updated detector preparer to set Hdf file output path.
The fluorescence detector to be used during an XML based scan now has
the Hdf writer set to output to the directory containing the GDA
produced Nexus files. The original path is restored after the scan
finishes.
Note the the Xspress3 implementation currently used by I18 needs to be
handled separated - DetectorHdfFunctions does not deal with
Xspress3WithFullCalculationsDetector (it is used only by I18).

Change-Id: I1d9df0623f27434f2d8be91118aa5b2c2f6634b5