K11-2141 Scan summary issues
- 'Points' in diffraction tomography:
Was 1 because for TWO_DIMENSION_LINE we assumed that the first axis in
trajectory would give us the result. For diffraction tomography X is the
first, and for a point that will be 1. We need to take the maximum from
both axes (in this case theta gives us the meaningful value). This is
true more generically for static points and 1D lines too, so their logic
changed too.

- The other issues mentioned in the ticket related to 'Step size' for
various shapes. This has been improved and simplified: static points and
2D points don't produce that string; everything else uses the same logic

Change-Id: I9002a1577a9bef49f06a9d0923b2bc99f147ce69