I11 (9.33) Switch mythen3 collection strategy to a compositional design
The Mythen3DetectorCollectionStrategy sub-class of SimpleAquire was
never tested, and sets the ImageMode to continuous, rather than
multiple, so scans never complete.

Instead, compose the required behaviour using the SoftwareStartStop
collection strategy, adding a MultipleImageModeDecorator and a
ConfigureAcquireTimeDecorator.

Also set lazyOpen on mythen3_hdf5, so it is possible to configure the
hdf writer before it has received it's first image. This fixes the
problem of GDA being unable to capture any images after the IOC is
restarted.

Change-Id: I20798453afd3937e61a581e99fd925b1d800d42b