I11-774 (9.33) Support per module bad channel lists
Automate the calculation of the bad channel list based on the modules
specified and their position in the datasets.

Also limit the bad channel list to the channels selected by the raw data
limits.

Move loading the modules file to before loading the bad channels files,
so that the module information can be used to determine which bad
channel files are loaded.

Add some notes on the tests I ran.

Also, add venv to .gitignore, since it is needed to run reduction.py but
shouldn't be in the repo.

Change-Id: Ie28680cb6210f5bb170fe2082ab41907d6f27f98