I11-774 (9.33) Support the supplied flatfield file
The .hdf5 files written by area detector and the .h5 files supplied for
the flatfield don't have the same format, so modify load_raw_data
function to take a flatfield_not_data argument, and use it to determine
which data node to use.

Also pass flatfield_not_data through load_and_histogram, as only the
caller of this function knows if it needs a flatfield or data.

Finally, update testing notes with flatfield examples.

Change-Id: Ief7d971b733677e40ca90faa548c75e8cbc96bda