K11-2175 Fix wrong scan being loaded
This was due to ShapeControls::updateControls being invoked both when a
scan of different shape to current selection was loaded AND when a shape
was manually changed. The function unconditionally mutated the axes of
the scan, swapping them for the cached axes corresponding to the new
shape selection. This is the intended behaviour for when the shape is
changed, but not when a scan is loaded. The fix in this change is coarse
but effective: a flag wraps #reload (invoked on scan load), and the
mutation is skipped when the flag is set.

Change-Id: I9f8509a4accaec99bacc46ae65533ca39fcf3115