I09 - Update dummy DCM energy speed to match PGM energy
Currently, using dcmenergy in dummy mode when moving will finish
instantly. This configuration updates it so it actually takes time to
move the motor position to better simulate live. DCM energy is in keV,
this reduces the speed of the motor by a factor of 1000 to match
pgmenergy eV speed.

Also added upper and lower limits.

Change-Id: Iced820a6dd8d5e94e1267cb6f9c0de2f3af414ca