I20 : QXes script updates after testing
- Minor tidying up and some refactoring to move scan preparation
functions to 'helper' classes.
- New 'prepare_run_qxes_scan function'. This can be used to run scans
using 1 or 2 rows of the spectrometer, by using optional energy start
and step parameters. It uses the Malcolm block that moves both rows
simultaneously - the gain and offset control the relative pitch between
the 2 rows, and upper, lower rows are enabled/disabled as required to
run the 1 or 2 colour scan with 1 or 2 rows.
- The upper and lower spectrometer row 'enabled' state has been added to
the processed data.

Change-Id: I875bfd8ca7b858c9dfb7e2cc602b4f14085315da