K11: Let gas_shaving_shutter's exceptions bubble up into main thread
At some point, an instruction to open the shutter before a scan failed
but the scan continued (collecting without beam). The problem was clear:
exceptions were being thrown in a background thread. The question then
becomes where should we throw? moveTo makes the most conceptual sense
(isBusy is not failing, neither is waitWhileBusy), though the maintainer
should be aware that moveTo does not get called during a scan. This
scannable is not used in scans in K11 so this solution works for them.

Change-Id: I55f996fabe0eccb8e606780dfe4539700d4a5228