I09-550 - Add back basic nixswr scannable
For backwards compatibility for beamline scripts, add back the original
nixswr DisplayEpicsPVClass.

Due to renaming nixswr to nixswr_repeat and nixswr_time, there was
confusion with beamline staff as their previous scripts no longer
worked. This makes it so that the new changes are still added while the
old way still works.

Change-Id: I96a9920a417a1936dc8ed22a9bf1b2530fc26684