I07 - fix datetime name collision bug, merge id calibration table
The beamline have had problems with scripts which 'from datetime import
datetime' because they have some other scripts which will import a
different datetime and overwrite it.  This change adds a distinct local
name for each use of datetime so the collision will not happen.  Also
merged their additions to the id gap calibration table and cleaned up
some imports

Change-Id: Ifd30bf607bd3257faf46679e1e40547a45f3eafe