I09-612 - Add dynamic metadata for I and J branch
Created DynamicScanMetadata class which is added as a scan listener.
Checks to see if a matching scannable is involved in a scan and
dynamically enables / disables metadata associated with it.

Have also switched beamline to new metadata configuration. Device names
are given to CommonBeamlineDevicesConfiguration. Some device names are
nexus devices, majority will be automatically wrapped by framework by
ConfiguredScannableNexusDevice or DefaultScannableNexusDevice. I and J
branches is added by DynamicScanMetadata class so that configuration is
only managed in once place.

Change-Id: I6b8c5747ecce867ad24e01a1118893c67a08725e