I20-658 : Make automatic XES peak finding work for all colour modes
- Automatic XES peak finding now works correctly when using '2 colour'
and '1 colour (both rows)' mode. The peak emission energy for each row is
found separately and used as the XES energy position during the main data
collection scan.
- In the case of '1 colour (both rows)', it is treated as '2 colour' (using
row 1 energy for the initial energy of row 2), so that the peak emission
can be found and used for each row.
- 'xes_peak_fit.py' has been updated to facilitate testing in dummy mode so
that each row has a different peak position.
- Unit tests have been updated.

Change-Id: I5ba3c60c8961115a10a53a778dc2a1ebe01c16af