I08-406 Update Sparse XANES script to change file path
With the new updates, the new scan file path would be the following:
.../2025/{visit_directory}/nexus/sparse_xanes/i08-12_001.nxs

I added sparse_xanes to the full path, so they are stored in a separate
folder. Also, for each energy, the script now appends _000, _001, _002,
etc.

The all_nexus_file_names field is also appending the file paths with the
same format.

Also, there was a comment that the last scan should scan all y positions
but this was not the case in practice, I have added a new else statement
to check this and to make random y positions be all y positions.

Change-Id: I5d9e17b22b2d756fe42ddcd1a20d71bfacfefe4a