I07-786:  Add methods to restart MALCOLM detector IOCs
This provides methods to automate something the beamline scientists say
they have to do every few days anyway.  It is hoped the problem will
eventually be fixed in MALCOLM and this will no longer be necessary but
for the mean time we have this workaround.

There seems to be no way to verify that an IOC has restarted, each
provides a status PV but this switches to running before the IOC is
actually finished.  This implementation uses time limits instead, based
on timing how long the IOCs take to restart and adding a margin.  This
is not foolproof but has been agreed with beamline scientists, they will
experiment with the methods and report back if the timings need to be
changed.

Change-Id: I5f9d3d2a0f682ae4b7213b8b6eafa894b497c262