I16-898, I16-969:  correct transformation metadata and geometry files
This has involved a great simplification of the geometry scannable as
well as replacing hard-coded transformation values in metadata with
corrected ones or scannable fields.  We also had to correct the path of
the geometry configuration files, this has been fixed before but seems
to have been lost in an update.

Change-Id: I4c6d0d3b04334c5e9c2db4e1057d65a91127f519