I09-2 - Fix LiveControlImage being disposed
When interacting with live controls and you hit the button to toggle the
stop button, change widget layout or display increment, the
LiveControlImage fails and can no longer display the image as the image
is disposed.

As the image is managed by SWTResourceManager, we don't need to dispose
of the image as this will be done for us which caused this bug.

Change-Id: I3c36c1ea1cd894ce5b326c65ab057482cfe01fcd